Cast time_t to unsigned long long to avoid compiler warning

+++ src/stats.c
@@ -360,7 +360,7 @@ void stats_send_statistics(stats_session_t *session) {
 	buffer_produced(response,
 		snprintf((char *)buffer_tail(response), buffer_spacecount(response),
 		"global last_reload timestamp %" PRIu64 "\n",
-		session->server->last_reload));
+		(unsigned long long)session->server->last_reload));
 
 	buffer_produced(response,
 		snprintf((char *)buffer_tail(response), buffer_spacecount(response),
